requirements.txt
PyGithub==1.59.1
requests==2.31.0
python-dotenv==1.0.0
